refactor: tools (#512)
